Add community, bottom navigation, and spiritual path improvements
Adds community screen, verbum bottom navigation widget, tests, and
various updates to spiritual paths, notifications, and services.
Removes tracked release keystore and ignores keystore files going forward.

+++ lib/models/spiritual_path.dart
@@ -63,8 +63,9 @@ class SpiritualPathProgress {
     this.lastCompletedAt,
   });
 
-  double progressFor(int totalDays) =>
-      totalDays == 0 ? 0 : (completedDays.length / totalDays).clamp(0, 1);
+  double progressFor(int totalDays) => totalDays == 0
+      ? 0
+      : (completedDays.length / totalDays).clamp(0, 1).toDouble();
 
   int nextDay(int totalDays) {
     for (var day = 1; day <= totalDays; day++) {
